nd_blk: fix theoretical support for >4GB apertures
If an aperture is larger than 4GB then nd_acpi_blk_region_do_io will
fail to find any length to transfer.

+++ drivers/block/nd/acpi_nfit.h
@@ -268,7 +268,7 @@ struct acpi_nfit_desc {
 	struct device *dev;
 	unsigned long dimm_dsm_force_en;
 	int (*blk_do_io)(struct nd_blk_region *ndbr, void *iobuf,
-			unsigned int len, int write, resource_size_t dpa);
+			u64 len, int write, resource_size_t dpa);
 };
 
 enum nd_blk_mmio_selector {
